+++ src/main/java/dk/kb/datahandler/oai/OaiHarvestClient.java
@@ -12,6 +12,7 @@ import java.net.http.HttpResponse.BodyHandlers;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.Base64;
+import java.util.List;
 
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
@@ -29,6 +30,7 @@ import org.xml.sax.InputSource;
 
 import dk.kb.datahandler.model.v1.JobDto;
 import dk.kb.datahandler.model.v1.OaiTargetDto;
+import dk.kb.datahandler.util.XmlUtils;
 import dk.kb.util.xml.XMLEscapeSanitiser;
 
 import org.w3c.dom.ls.*;
@@ -78,7 +80,7 @@ public class OaiHarvestClient {
 
         String errorMessage = getErrorMessage(document);
          if (errorMessage != null && errorMessage.trim().length() >1) {                       
-            log.info("Error message from OAI server when harvesting set:"+set +" message:"+errorMessage);                    
+            log.info("Error message from OAI server when harvesting set:"+set +" message:"+errorMessage);
             oaiResponse.setError(true);
             return oaiResponse;// will have no records
          }
@@ -130,15 +132,21 @@ public class OaiHarvestClient {
         return uri;
     }
 
-    public ArrayList<OaiRecord> extractRecordsFromXml(Document document) {
-        NodeList nList = document.getElementsByTagName("record"); 
+    public ArrayList<OaiRecord> extractRecordsFromXml( Document document ) {
+
+        NodeList nList = document.getElementsByTagName("ListRecords"); //This is the OAI tag, not part of preservica data.
+        Element recordsElement = (Element) nList.item(0); //always one.
+        List<Element> elementList = new ArrayList<Element>();
+        elementList.add(recordsElement);
+
+        List<Element> recordList = XmlUtils.getDirectChildrenByTag(elementList, "record");
 
         ArrayList<OaiRecord> records= new ArrayList<OaiRecord>();
-        for (int i =0;i<nList.getLength();i++) {                                         
+        for (int i =0; i<recordList.size(); i++) {
 
             OaiRecord oaiRecord = new OaiRecord();
             records.add(oaiRecord);
-            Element record =  (Element)nList.item(i);                                                                    
+            Element record =  (Element)recordList.get(i);
             String identifier =  record.getElementsByTagName("identifier").item(0).getTextContent();                        
             String datestamp =  record.getElementsByTagName("datestamp").item(0).getTextContent();
             String headerStatus = getHeaderStatus(record);            
@@ -344,7 +352,5 @@ public class OaiHarvestClient {
             return "?";                    
         }
     }
-
-
 
 }